# src/main/resources/swagger/swagger-input.yml
swagger: "2.0"
info:
  description: "PPM Patient Adapter"
  version: "1.0.0"
  title: "PPM Patient Adapter"
tags:
- name: "admin"
  description: "Everything about patient"
schemes:
- "https"
- "http"
paths:
  v1/profiles:
    post:
      tags:
      - "patient"
      summary: "check deleted profiles"
      description: "check deleted"
      operationId: "checkDeletedProfiles"
      consumes:
      - "application/json; text/plain; charset=utf-8"
      produces:
      - "application/json; text/plain; charset=utf-8"
      parameters:
      - in: "body"
        name: "patientIds"
        required: true
        description: "list id of patient profiles"
        schema:
          type: "array"
          items:
            type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              type: "string"
        400:
          description: "Invalid input"
        500:
          description: "Internal server error"

  v1/patients:
    post:
      tags:
      - "patient"
      summary: "add patients"
      description: "add patients by list of id patients"
      operationId: "addPatientProfiles"
      consumes:
      - "application/json; text/plain; charset=utf-8"
      produces:
      - "application/json; text/plain; charset=utf-8"
      parameters:
      - in: "body"
        name: "patientIds"
        required: true
        description: "list id of patient profiles"
        schema:
          type: "array"
          items:
            type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              type: "string"
        400:
          description: "Invalid input"
        500:
          description: "Internal server error"

    delete:
      tags:
      - "patient"
      summary: "delete patients"
      description: "delete patiens by ids of patients"
      operationId: "deletePatientProfiles"
      consumes:
      - "application/json; text/plain; charset=utf-8"
      parameters:
      - in: "body"
        name: "patientIds"
        required: true
        description: "list id of patient profiles"
        schema:
          type: "array"
          items:
            type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "string"
        400:
          description: "Invalid input"
        500:
          description: "Internal server error"